removing TODOs and movign AilaGeneratorService interface to interfaces from ailaservices
