+++ packages/aila/src/core/AilaServices.ts
@@ -63,23 +63,23 @@ export interface AilaChatService {
   startStreaming(abortController?: AbortController): ReadableStream;
 }
 
-export interface AilaQuizService {
-  generateMathsExitQuizPatch(
-    lessonPlan: LooseLessonPlan,
-  ): Promise<JsonPatchDocument>;
-}
-// TODO: GCLOMAX - move this to interfaces and rename.
-export interface AilaQuizGeneratorService {
-  generateMathsExitQuizPatch(
-    lessonPlan: LooseLessonPlan,
-    relevantLessons?: AilaRagRelevantLesson[],
-  ): Promise<Quiz[]>;
-  generateMathsStarterQuizPatch(
-    lessonPlan: LooseLessonPlan,
-    relevantLessons?: AilaRagRelevantLesson[],
-  ): Promise<Quiz[]>;
-  // invoke(lessonPlan: LooseLessonPlan): Promise<Quiz[]>;
-}
+// export interface AilaQuizService {
+//   generateMathsExitQuizPatch(
+//     lessonPlan: LooseLessonPlan,
+//   ): Promise<JsonPatchDocument>;
+// }
+// // TODO: GCLOMAX - move this to interfaces and rename.
+// export interface AilaQuizGeneratorService {
+//   generateMathsExitQuizPatch(
+//     lessonPlan: LooseLessonPlan,
+//     relevantLessons?: AilaRagRelevantLesson[],
+//   ): Promise<Quiz[]>;
+//   generateMathsStarterQuizPatch(
+//     lessonPlan: LooseLessonPlan,
+//     relevantLessons?: AilaRagRelevantLesson[],
+//   ): Promise<Quiz[]>;
+//   // invoke(lessonPlan: LooseLessonPlan): Promise<Quiz[]>;
+// }
 
 export interface AilaServices {
   readonly userId: string | undefined;
